fix 100cb fill fails, always path hint key 8
