+++ randomizer/Lists/Exceptions.py
@@ -95,3 +95,9 @@ class BossOutOfLocationsException(FillException):
     """Exception triggered when there are no valid levels to put a boss."""
 
     pass
+
+
+class CBFillFailureException(FillException):
+    """Exception triggered when CB rando fails to correctly generate a valid set of groups."""
+
+    pass
