+++ randomizer/LogicClasses.py
@@ -198,3 +198,56 @@ class Sphere:
         self.seedBeaten = False
         self.availableGBs = 0
         self.locations = []
+
+
+class ColoredBananaGroup:
+    """Stores data for each group of colored bananas."""
+
+    def __init__(self, *, group=0, name="No Location", map_id=0, konglist=[], region=None, logic=None, vanilla=False, locations=[]):
+        """Initialize with given parameters."""
+        self.group = group
+        self.name = name
+        self.map = map_id
+        self.kongs = konglist
+        self.locations = locations  # 5 numbers: {int amount, float scale, int x, y, z}
+        self.region = region
+        if logic is None:
+            self.logic = lambda l: True
+        else:
+            self.logic = logic
+        self.selected = False
+
+
+class Balloon:
+    """Stores data for each balloon."""
+
+    def __init__(self, *, id=0, name="No Location", map_id=0, speed=0, konglist=[], region=None, logic=None, vanilla=False, points=[]):
+        """Initialize with given parameters."""
+        self.id = id
+        self.name = name
+        self.map = map_id
+        self.speed = speed
+        self.kongs = konglist
+        self.points = points  # 3 numbers: [int x, y, z]
+        self.region = region
+        if logic is None:
+            self.logic = lambda l: True
+        else:
+            self.logic = logic
+        self.spawnPoint = self.setSpawnPoint(points)
+        self.selected = False
+
+    def setSpawnPoint(self, points=[]):
+        """Set the spawn point of a balloon based on its path."""
+        spawnX = 0
+        spawnY = 0
+        spawnZ = 0
+        for p in points:
+            spawnX += p[0]
+            spawnY += p[1]
+            spawnZ += p[2]
+        spawnX /= len(points)
+        spawnY /= len(points)
+        spawnY -= 100  # Most balloons are at least 100 units off the ground
+        spawnZ /= len(points)
+        return [int(spawnX), int(spawnY), int(spawnZ)]
