+++ tests/test_spoiler.py
@@ -132,7 +132,8 @@ def generate_lo_rando_race_settings():
     data["microhints_enabled"] = "base"  # off/base/all
     data["smaller_shops"] = True  # likely to be True in item rando, many settings force it to be false
     data["alter_switch_allocation"] = False  # likely to be True, easier to test things when false
-    data["random_starting_region"] = True  # likely to be False
+    data["random_starting_region"] = False  # likely to be False
+    data["random_fairies"] = False  # likely to be False
 
     return data
